Put benchmark in function_benchmark.cc
